requirements.txt
APScheduler==3.6.3
astroid==2.9.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
et-xmlfile==1.1.0
google-api-core==2.7.1
google-api-python-client==2.29.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
networkx==2.6.3
numpy==1.22.3
openpyxl==3.0.9
platformdirs==2.5.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycel==1.0b30
pylint==2.12.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
toml==0.10.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.13.3
